backend + web/smartfruit/requirements.txt
Django==5.1.6
djangorestframework==3.16.0
scikit-learn==1.5.1
numpy==2.1.1
pandas>=2.0.0
joblib>=1.3.0
requests==2.32.3
gunicorn==23.0.0
influxdb-client==1.36.0
python-dotenv